RuntimeException -> SparkIllegalArgumentException
